requirements.txt
Django==4.2.14
djangorestframework==3.14.0
psycopg2-binary==2.9.6
drf-spectacular==0.26.4
djangorestframework-simplejwt==5.5.0